scripts/unsplash_requirements.txt
pyperclip==1.8.2
python-dotenv==0.19.2
